refactor: extract fsw.py from networks/_legacy_full (4/N)
