+++ stisim/networks/__init__.py
@@ -1,4 +1,5 @@
 """STIsim networks package — split across base, mf, fsw, msm, layered_networks, matchers."""
 from .base import NoPartnersFound, BasePars, NetworkPars, BaseNetwork
 from .mf import MFPars, MFNetwork
+from .fsw import SWPars, SWNetwork
 from ._legacy_full import *  # noqa: F401, F403 (remaining classes during refactor)
